Implemented almost everything

+++ src/main/java/cz/cuni/mff/pijalekj/entities/EntityStats.java
@@ -1,10 +1,10 @@
 package cz.cuni.mff.pijalekj.entities;
 
 public class EntityStats {
-    public long credits;
+    public int credits;
     public int[] ownedGoods;
 
-    public EntityStats(long credits, int[] ownedGoods) {
+    public EntityStats(int credits, int[] ownedGoods) {
         this.credits = credits;
         this.ownedGoods = ownedGoods;
     }
